# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Email Microservice API",
    "description": "API untuk Email Microservice yang mendukung pengiriman email, manajemen template, dan pelacakan email.",
    "version": "1.0.0",
    "contact": {
      "name": "Tim API Email Microservice",
      "email": "api-support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development Server"
    },
    {
      "url": "https://api.example.com/api/v1",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/api-keys": {
      "post": {
        "tags": ["API Key"],
        "summary": "Membuat API key baru",
        "description": "Membuat API key baru untuk mengakses API",
        "operationId": "createApiKey",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ],
        "requestBody": {
          "description": "Payload pembuatan API key",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiKeyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key berhasil dibuat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autentikasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["API Key"],
        "summary": "Mendapatkan daftar API key",
        "description": "Mendapatkan daftar API key yang tersedia",
        "operationId": "getApiKeys",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Nomor halaman",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Jumlah item per halaman",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "active",
            "in": "query",
            "description": "Filter berdasarkan status aktif",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar API key berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeysListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autentikasi gagal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ApiKeyRequest": {
        "type": "object",
        "required": ["name", "permissions"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nama untuk API key"
          },
          "description": {
            "type": "string",
            "description": "Deskripsi API key"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["send_email", "manage_templates", "view_tracking"]
            },
            "description": "Permission untuk API key"
          },
          "rateLimits": {
            "type": "object",
            "properties": {
              "maxRequestsPerMinute": {
                "type": "integer",
                "description": "Jumlah maksimum request per menit"
              },
              "maxEmailsPerDay": {
                "type": "integer",
                "description": "Jumlah maksimum email per hari"
              }
            },
            "description": "Batas penggunaan"
          },
          "expiry": {
            "type": "string",
            "format": "date-time",
            "description": "Tanggal kedaluwarsa"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Metadata tambahan"
          }
        }
      },
      "ApiKeyResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "ID API key"
              },
              "key": {
                "type": "string",
                "description": "API key value (hanya ditampilkan sekali)"
              },
              "name": {
                "type": "string",
                "description": "Nama API key"
              },
              "permissions": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Permission API key"
              },
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "description": "Waktu pembuatan"
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time",
                "description": "Waktu kedaluwarsa"
              }
            }
          }
        }
      },
      "ApiKeysListResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["success"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "apiKeys": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "ID API key"
                    },
                    "name": {
                      "type": "string",
                      "description": "Nama API key"
                    },
                    "description": {
                      "type": "string",
                      "description": "Deskripsi API key"
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "Permission API key"
                    },
                    "active": {
                      "type": "boolean",
                      "description": "Status API key"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Waktu pembuatan"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Waktu kedaluwarsa"
                    },
                    "lastUsedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Waktu penggunaan terakhir"
                    }
                  }
                }
              },
              "pagination": {
                "type": "object",
                "properties": {
                  "currentPage": {
                    "type": "integer",
                    "description": "Halaman saat ini"
                  },
                  "totalPages": {
                    "type": "integer",
                    "description": "Total halaman"
                  },
                  "totalItems": {
                    "type": "integer",
                    "description": "Total item"
                  },
                  "itemsPerPage": {
                    "type": "integer",
                    "description": "Jumlah item per halaman"
                  }
                }
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["error"]
          },
          "traceID": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "properties": {
              "errors": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string",
                      "description": "Nama field yang error"
                    },
                    "message": {
                      "type": "string",
                      "description": "Deskripsi error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "AdminTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      }
    }
  }
}